meedere windows openen+windows slepen

+++ Tablr/src/TablesWindow.java
@@ -1,12 +1,11 @@
-import java.awt.Frame;
 import java.awt.Graphics;
 import java.awt.Color;
 import java.awt.Font;
 
 public class TablesWindow extends SubWindow {
 
-    public TablesWindow(TableManager tableManager){
-        super(tableManager);
+    public TablesWindow(TableManager tableManager, int xcor, int ycor, int width, int height) {
+        super(tableManager, xcor, ycor, width, height);
     }
 
     /**
@@ -17,42 +16,52 @@ public class TablesWindow extends SubWindow {
      * Single-click: Selects the chosen table
      */
     @Override
-    public void handleMouseEvent(Frame frame, CanvasWindow window, int x, int y, int clickCount) {
-        Table currentlySelectedTable = tableManager.getSelectedTable();
-        int idx = getIdx1D(frame.getHeight() - 30,tableManager.getMaxTablePerCol(),frame.getWidth(),tableManager.getMaxTablePerRow(),x,y,0,0);
-        if(clickCount==1){
-            tableManager.selectTable(idx);
-            tableManager.saveNewName();
-        }
-        //mouse double clicked
-        else if(clickCount==2){
-            
-            int rows = (int) Math.ceil((double) tableManager.getTables().size() / tableManager.getMaxTablePerRow());
-            int clickableHeight = rows * (frame.getHeight() - 30) / tableManager.getMaxTablePerCol(); // Nog niet volledig correct -> Bij meer dan 1 tabel per rij, kunnen de laatste tabellen niet toegevoegd worden
-            //create new table list entry
-            if(y>clickableHeight){
-                tableManager.createAndAddTable();
-            }
+    public void handleMouseEvent(int id,int x, int y, int clickCount) {
+        SubWindowManager.switchSubWindow(x,y, id);
+        SubWindow curWindow = SubWindowManager.getWindow();
+        if (this.equals(curWindow) && cursorInside(x, y)) {
+            if(id==500) {
 
-            //open table design mode or rows mode for table
-            else if (tableManager.isValidTable(currentlySelectedTable)) {
-                Table selected = tableManager.getSelectedTable();
-                if(selected!=null) {
+                Table currentlySelectedTable = tableManager.getSelectedTable();
+                int idx = getIdx1D(height - 2 * borderThickness, tableManager.getMaxTablePerCol(), width-2*borderThickness, tableManager.getMaxTablePerRow(), x, y, xcor + borderThickness, ycor + borderThickness);
+                if (clickCount == 1) {
+                    tableManager.selectTable(idx);
+                    tableManager.saveNewName();
+                }
+                //mouse double clicked
+                else if (clickCount == 2) {
 
-                    //if table has no columns, change to design mode
-                    if (selected.getCols().isEmpty()) {
-                        SubWindowManager.toTableDesignWindow(tableManager);
-                        window.setTitle(CONST_TABLE_COLUMN_MODE_TITLE + " - " + selected.getName());
+                    int rows = (int) Math.ceil((double) tableManager.getTables().size() / tableManager.getMaxTablePerRow());
+                    int clickableHeight = ycor+rows * (height - 30 - 2*borderThickness) / tableManager.getMaxTablePerCol(); // Nog niet volledig correct -> Bij meer dan 1 tabel per rij, kunnen de laatste tabellen niet toegevoegd worden
+                    //create new table list entry
+                    if (y > clickableHeight) {
+                        tableManager.createAndAddTable();
                     }
 
-                    //else change to rows mode
-                    else {
-                        SubWindowManager.toTableRowsWindow(tableManager);
-                        window.setTitle(CONST_TABLE_ROW_MODE_TITLE + " - " + selected.getName());
+                    //open table design mode or rows mode for table
+                    else if (tableManager.isValidTable(currentlySelectedTable)) {
+                        Table selected = tableManager.getSelectedTable();
+                        if (selected != null) {
+
+                            //if table has no columns, change to design mode
+                            if (selected.getCols().isEmpty()) {
+                                SubWindowManager.toTableDesignWindow(tableManager);
+                            }
+
+                            //else change to rows mode
+                            else {
+                                SubWindowManager.toTableRowsWindow(tableManager);
+                            }
+                        }
                     }
                 }
+            }else if(id==506){
+                drag(x,y);
             }
+
+
         }
+
     }
 
     /**
@@ -65,27 +74,31 @@ public class TablesWindow extends SubWindow {
      * Any non-special character: appends the character to the name
      */
     @Override
-    public void handleKeyEvent(CanvasWindow window, int keyCode, char keyChar, boolean isControlDown) {
-        //del key
-        if (keyCode == 127) {
-            tableManager.deleteTable();
-        }
+    public void handleKeyEvent(int id,int keyCode, char keyChar, boolean isControlDown) {
+        if (this.equals(SubWindowManager.getWindow())) {
+            if (id == 401) {
+                //del key
+                if (keyCode == 127) {
+                    tableManager.deleteTable();
+                }
 
-        //escape key
-        else if(keyCode==27){
-            tableManager.undoNameEdits();
-            tableManager.unselectTable();
-        }
+                //escape key
+                else if (keyCode == 27) {
+                    tableManager.undoNameEdits();
+                    tableManager.unselectTable();
+                }
 
-        //enter key
-        else if(keyCode==10){
-            tableManager.saveNewName();
-            tableManager.unselectTable();
-        }
+                //enter key
+                else if (keyCode == 10) {
+                    tableManager.saveNewName();
+                    tableManager.unselectTable();
+                }
 
-        //character keys
-        else if(!(keyCode>=16 && keyCode<=20)) {
-            tableManager.editTableName(keyChar);
+                //character keys
+                else if (!(keyCode >= 16 && keyCode <= 20)) {
+                    tableManager.editTableName(keyChar);
+                }
+            }
         }
     }
 
@@ -94,21 +107,20 @@ public class TablesWindow extends SubWindow {
      * If a name is invalid: mark it as red otherwise keep it gray
      */
     @Override
-    public void drawMode(Frame frame, Graphics g){
-        int width = frame.getWidth();
-        int height = frame.getHeight() - 30; // 30 is the height of the Title bar
+    public void drawMode(Graphics g){
+        int height_ = height - 30; // 30 is the height of the Title bar
 
-        int entryWidth = width/tableManager.getMaxTablePerRow();
-        int entryHeight = height/tableManager.getMaxTablePerCol();
+        int entryWidth = (width-2*borderThickness)/tableManager.getMaxTablePerRow();
+        int entryHeight = (height-2*borderThickness)/tableManager.getMaxTablePerCol();
         g.setColor(Color.gray);
-        g.fillRect(0, 0, width, height);
+        g.fillRect(xcor, ycor, width, height_);
 
         int row = 0;
         int col = 0;
 
         for(Table table:tableManager.getTables()) {
             String name = table.getName();
-                      
+
             if(table.equals(tableManager.getSelectedTable())) {
                 name += "\uD83D\uDC46";
             }
@@ -119,13 +131,13 @@ public class TablesWindow extends SubWindow {
             }else{
                 g.setColor(Color.lightGray);
             }
-            g.fillRect(col*entryWidth,row*entryHeight,entryWidth,entryHeight);
+            g.fillRect(xcor+borderThickness+col*entryWidth,ycor+borderThickness+row*entryHeight,entryWidth,entryHeight);
             g.setColor(Color.darkGray);
-            g.drawRect(col*entryWidth,row*entryHeight, entryWidth, entryHeight);
+            g.drawRect(xcor+borderThickness+col*entryWidth,ycor+borderThickness+row*entryHeight, entryWidth, entryHeight);
 
             g.setColor(Color.black);
             g.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 14));
-            g.drawString(name, col * entryWidth + 10, row * entryHeight + (entryHeight / 2) + 5);
+            g.drawString(name, xcor+borderThickness+col * entryWidth + 10, ycor+borderThickness+row * entryHeight + (entryHeight / 2) + 5);
 
 
             col++;
@@ -134,6 +146,11 @@ public class TablesWindow extends SubWindow {
                 col = 0;
             }
         }
+
+        drawBorders(g);
+        g.setColor(Color.black);
+        g.drawString("TablesWindow",xcor,ycor+10);
+        drawCloseButton(g);
     }
 
 }
